Fix #102, rename files according to convention
Updates the HS file naming to be consistent with the documented
recommendations.

This also removes "hs_custom.h" as it was no longer used (remnant).
