+++ fsw/src/hs_cmds.c
@@ -29,7 +29,7 @@
 #include "hs_cmds.h"
 #include "hs_monitors.h"
 #include "hs_msgids.h"
-#include "hs_events.h"
+#include "hs_eventids.h"
 #include "hs_utils.h"
 #include "hs_version.h"
