+++ fsw/src/hs_utils.c
@@ -27,10 +27,9 @@
 *************************************************************************/
 #include "hs_app.h"
 #include "hs_utils.h"
-#include "hs_custom.h"
 #include "hs_monitors.h"
 #include "hs_msgids.h"
-#include "hs_events.h"
+#include "hs_eventids.h"
 #include "hs_version.h"
 
 /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
